+++ CHANGES.md
@@ -1,4 +1,6 @@
 #### [unreleased]
+
+#### 2.1.0 / 2020-07-20
 * update block build process
 * add title and description attributes for a11y
 * add PDF svg for block